IPs - production modification

+++ src/Controller/IpsController.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use Cake\I18n\FrozenTime;
+
 /**
  * Ips Controller
  *
@@ -18,8 +20,23 @@ class IpsController extends AppController
      */
     public function index()
     {
+        $customer_id = $this->request->getParam('customer_id');
+        $this->set('customer_id', $customer_id);
+        
+        $contract_id = $this->request->getParam('contract_id');
+        $this->set('contract_id', $contract_id);
+
+        $conditions = [];
+        if (isset($customer_id)) {
+            $conditions += ['Ips.customer_id' => $customer_id];
+        }
+        if (isset($contract_id)) {
+            $conditions += ['Ips.contract_id' => $contract_id];
+        }
+
         $this->paginate = [
             'contain' => ['Customers', 'Contracts'],
+            'conditions' => $conditions,
         ];
         $ips = $this->paginate($this->Ips);
 
@@ -49,7 +66,23 @@ class IpsController extends AppController
      */
     public function add()
     {
+        $customer_id = $this->request->getParam('customer_id');
+        $this->set('customer_id', $customer_id);
+        
+        $contract_id = $this->request->getParam('contract_id');
+        $this->set('contract_id', $contract_id);
+
         $ip = $this->Ips->newEmptyEntity();
+
+        $conditions = [];
+        if (isset($customer_id)) {
+            $ip = $this->Ips->patchEntity($ip, ['customer_id' => $customer_id]);
+            $conditions += ['customer_id' => $customer_id];
+        }
+        if (isset($contract_id)) {
+            $ip = $this->Ips->patchEntity($ip, ['contract_id' => $contract_id]);
+        }
+
         if ($this->request->is('post')) {
             $ip = $this->Ips->patchEntity($ip, $this->request->getData());
             if ($this->Ips->save($ip)) {
@@ -60,7 +93,7 @@ class IpsController extends AppController
             $this->Flash->error(__('The ip could not be saved. Please, try again.'));
         }
         $customers = $this->Ips->Customers->find('list', ['order' => ['company', 'first_name', 'last_name']]);
-        $contracts = $this->Ips->Contracts->find('list', ['order' => 'number']);
+        $contracts = $this->Ips->Contracts->find('list', ['order' => 'number', 'conditions' => $conditions]);
         $this->set(compact('ip', 'customers', 'contracts'));
     }
 
@@ -73,9 +106,21 @@ class IpsController extends AppController
      */
     public function edit($id = null)
     {
+        $customer_id = $this->request->getParam('customer_id');
+        $this->set('customer_id', $customer_id);
+        
+        $contract_id = $this->request->getParam('contract_id');
+        $this->set('contract_id', $contract_id);
+
         $ip = $this->Ips->get($id, [
             'contain' => [],
         ]);
+
+        $conditions = [];
+        if (isset($customer_id)) {
+            $conditions += ['customer_id' => $customer_id];
+        }
+        
         if ($this->request->is(['patch', 'post', 'put'])) {
             $ip = $this->Ips->patchEntity($ip, $this->request->getData());
             if ($this->Ips->save($ip)) {
@@ -86,7 +131,7 @@ class IpsController extends AppController
             $this->Flash->error(__('The ip could not be saved. Please, try again.'));
         }
         $customers = $this->Ips->Customers->find('list', ['order' => ['company', 'first_name', 'last_name']]);
-        $contracts = $this->Ips->Contracts->find('list', ['order' => 'number']);
+        $contracts = $this->Ips->Contracts->find('list', ['order' => 'number', 'conditions' => $conditions]);
         $this->set(compact('ip', 'customers', 'contracts'));
     }
 
@@ -101,12 +146,40 @@ class IpsController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $ip = $this->Ips->get($id);
-        if ($this->Ips->delete($ip)) {
-            $this->Flash->success(__('The ip has been deleted.'));
-        } else {
-            $this->Flash->error(__('The ip could not be deleted. Please, try again.'));
+        
+        if ($this->addToRemovedIps($id))
+        {
+            if ($this->Ips->delete($ip)) {
+                $this->Flash->success(__('The ip has been deleted.'));
+            } else {
+                $this->Flash->error(__('The ip could not be deleted. Please, try again.'));
+            }
         }
 
         return $this->redirect(['action' => 'index']);
+    }
+    
+    private function addToRemovedIps ($id = null)
+    {
+        $ip = $this->Ips->get($id);
+        
+        $this->RemovedIps = $this->getTableLocator()->get('RemovedIps');
+        
+        $removedIp = $this->RemovedIps->newEmptyEntity();
+        $removedIp = $this->RemovedIps->patchEntity($removedIp, $ip->toArray());
+        
+        // TODO - add who and when deleted this
+        $removedIp->removed = FrozenTime::now();
+        $removedIp->removed_by = $_SESSION['login_id'];
+        
+        if ($this->RemovedIps->save($removedIp)) {
+            $this->Flash->success(__('The removed ip has been saved.'));
+            return true;
+        }
+        
+        debug($removedIp);
+
+        $this->Flash->error(__('The removed ip could not be saved. Please, try again.'));
+        return false;
     }
 }
